Move getTTLFromAnnotations to a common file
