+++ source/source.go
@@ -16,7 +16,11 @@ limitations under the License.
 
 package source
 
-import "github.com/kubernetes-incubator/external-dns/endpoint"
+import (
+	"github.com/kubernetes-incubator/external-dns/endpoint"
+	"strconv"
+	log "github.com/Sirupsen/logrus"
+)
 
 const (
 	// The annotation used for figuring out which controller is responsible
@@ -34,4 +38,21 @@ const (
 // Source defines the interface Endpoint sources should implement.
 type Source interface {
 	Endpoints() ([]*endpoint.Endpoint, error)
+}
+
+func getTTLFromAnnotations(annotations map[string]string) endpoint.TTL {
+	ttlAnnotation, exists := annotations[ttlAnnotationKey]
+	if !exists {
+		return endpoint.TTL(0)
+	}
+	ttlValue, err := strconv.ParseInt(ttlAnnotation, 10, 64)
+	if err != nil {
+		log.Warnf("%v is not a valid TTL value", ttlAnnotation)
+		return endpoint.TTL(0)
+	}
+	if ttlValue < 0 {
+		log.Warnf("TTL must be a non-negative integer", ttlAnnotation)
+		return endpoint.TTL(0)
+	}
+	return endpoint.TTL(ttlValue)
 }
